% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz numbers game.}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{Any real number}
}
\value{
A real number, the word "Fizz", the word "Buzz", the word "FizzBuzz"
}
\description{
This function replicates the numbers game FizzBuzz, where you take turns
counting sequentially, replacing numbers divisible by 3 with Fizz,
numbers divisible by 5 with Buzz, and numbers divisible by both with
FizzBuzz.
}
\details{
The function will produce the word "Fizz" when a number is evenly
divisible by 3, and the word "Buzz" when a number is evenly divisible by 5.
When a number is evenly divisible by both 3 and 5 the function will
produce the word "FizzBuzz".
}
\examples{
FizzBuzz(20)
FizzBuzz(20-50)
}
